+++ cmd/autobump/python.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"errors"
+	"fmt"
 	"os"
 	"path/filepath"
 
@@ -19,6 +21,8 @@ type Python struct {
 	ProjectConfig ProjectConfig
 }
 
+var ErrPyprojectNotFound = errors.New("pyproject.toml not found")
+
 func (p Python) GetProjectName() (string, error) {
 	return getPyprojectName(p.ProjectConfig)
 }
@@ -28,13 +32,16 @@ func getPyprojectName(projectConfig ProjectConfig) (string, error) {
 
 	_, err := os.Stat(pyprojectTomlPath)
 	if err != nil {
-		return "", err
+		if os.IsNotExist(err) {
+			return "", ErrPyprojectNotFound
+		}
+		return "", fmt.Errorf("error checking pyproject.toml path: %w", err)
 	}
 
 	var pyProject PyProject
 	_, err = toml.DecodeFile(pyprojectTomlPath, &pyProject)
 	if err != nil {
-		return "", err
+		return "", fmt.Errorf("error decoding pyproject.toml: %w", err)
 	}
 
 	return pyProject.Project.Name, nil